feat: envelopeedit 로직 수정
