+++ Projects/Feature/SSEnvelope/Sources/SpecificEnvelopeEdit/SpecificEnvelopeType.swift
@@ -1,0 +1,14 @@
+//
+//  SpecificEnvelopeType\.swift
+//  SSEnvelope
+//
+//  Created by MaraMincho on 11/28/24.
+//  Copyright © 2024 com.oksusu. All rights reserved.
+//
+
+import Foundation
+
+public enum SpecificEnvelopeType: Equatable, Sendable {
+  case sent
+  case received
+}
